+++ src/disassembler/core/opcode.ts
@@ -1544,19 +1544,6 @@ export class Opcode {
 			valueName = (val >= 0) ? '+' : '';
 			valueName += val.toString();
 		}
-
-/* TODO: REMOVE
-		// E.g. "LD DE,nnnn"
-		else if (this.valueType == NumberType.NUMBER_BYTE) { 
-			// byte
-			valueName = Format.getHexFormattedString(this.value, 2);
-		}
-		else {
-			// word
-			valueName = Format.getHexFormattedString(this.value, 4);
-		}
-*/
-
 		else if (this.valueType == NumberType.NUMBER_BYTE) {
 			// byte
 			valueName = Format.getHexFormattedString(this.value, 2);
@@ -1568,7 +1555,6 @@ export class Opcode {
 			else
 				valueName = funcGetLabel(this.value);
 		}
-
 
 		// Normal disassembly
 		this.disassembledText = util.format(this.name, valueName);