+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 3.2.3
 - Fix for #109: Extension host terminated unexpectedly 3 times within the last 5 minutes. aduh95/graphviz.js has been replaced with node-graphviz.
+- Changed disassembly of e.g. 'LD DE,8000.1' to 'LD DE,$8000.1'
 
 # 3.2.2
 - Fix for #108: Reverse engineering: Call Graph and Flow Chart failing to render.